refactor(eslint-plugin-i18n): rename sort-i18n-message-keys to sort-message-locales
