+++ packages/eslint-plugin-i18n/README.md
@@ -31,4 +31,4 @@ Rules for i18n linting.
 | [restrict-i18n-imports](docs/rules/restrict-i18n-imports.md)                               | Prohibits importing text files (`i18n.ts`) from directories of other levels.                                                           |
 | [auto-generate-translation-message-id](docs/rules/auto-generate-translation-message-id.md) | Checks for the presence of an identifier in the translation object and adds it if missing                                              |
 | [multiline-i18n-meta-object](docs/rules/multiline-i18n-meta-object.md)                     | In `i18n.ts`, requires `meta` in `createMessages` entries to be multiline objects (with line breaks), not single-line `meta: { ... }`. |
-| [sort-i18n-message-keys](docs/rules/sort-i18n-message-keys.md)                             | Enforces key order in message objects: `ru`, `en`, then other locales (source order), then `meta`.                                     |
+| [sort-message-locales](docs/rules/sort-message-locales.md)                                 | Enforces locale order in message objects: `ru`, `en`, then other locales (source order), then `meta`.                                  |
