+++ packages/eslint-plugin-i18n/index.ts
@@ -2,7 +2,7 @@ import {rule as autoGenerateTranslationMessageId} from './rules/auto-generate-tr
 import {rule as detectIncorrectCalls} from './rules/detectIncorrectCalls';
 import {rule as multilineI18nMetaObject} from './rules/multiline-i18n-meta-object';
 import {rule as restrictI18nImports} from './rules/restrict-i18n-imports';
-import {rule as sortI18nMessageKeys} from './rules/sort-i18n-message-keys';
+import {rule as sortMessageLocales} from './rules/sort-message-locales';
 import {rule as stringLiteralKeys} from './rules/string-literal-keys';
 
 import pkg from './package.json';
@@ -17,7 +17,7 @@ export const rules = {
     'detect-incorrect-calls': detectIncorrectCalls,
     'multiline-i18n-meta-object': multilineI18nMetaObject,
     'restrict-i18n-imports': restrictI18nImports,
-    'sort-i18n-message-keys': sortI18nMessageKeys,
+    'sort-message-locales': sortMessageLocales,
     'string-literal-keys': stringLiteralKeys,
 };
